% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enter_bls_grades}
\alias{enter_bls_grades}
\title{Enter student grades in BLS marks spreadsheet}
\usage{
enter_bls_grades(
  grades_df,
  bls_marks_spreadsheet_filename,
  id_col = "id",
  grade_col = "grade"
)
}
\arguments{
\item{grades_df}{A data frame with student grades in NTU GBA abbreviated format like DHIGH, DMID, etc}

\item{bls_marks_spreadsheet_filename}{The filename of BLS marks spreadsheet, e.g. "PSYC40545_29120_M01_CWK_100-202425.xlsx"}

\item{id_col}{The column in `grades_df` that denotes the student ID (N number)}

\item{grade_col}{The column in `grades_df` with the students' grades}
}
\value{
TRUE if BLS marks spreadsheet is successfully written, FALSE otherwise
}
\description{
Enter student grades in BLS marks spreadsheet
}
